Fixed some issues with Call_Classes

+++ Call_Classes/getNonReturnersCall.java
@@ -23,7 +23,7 @@ public class getNonReturnersCall extends ClientCall implements Serializable {
 	/**
 	 * @return the days
 	 */
-	public long getDays() {
+	public int getDays() {
 		return days;
 	}
